+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Blip is the web front end for YourLoops system.
 It is based on Tidepool Blip 1.27.
 
 ## 3.0.2-beta12 - 2022-11-03
+### Added
+- YLP-1850 Patient search should comply to INS standard
 ### Engineering use
 - YLP-1624 Fix error messages when inviting a patient already in the team
 - YLP-1873 Fix patient profile update not visible in dashboard